Conflicts:
	src/main/java/edu/stanford/epad/epadws/handlers/admin/ServerStatusHandler.java
	src/main/java/edu/stanford/epad/epadws/handlers/core/EPADHandler.java
	src/main/java/edu/stanford/epad/epadws/handlers/dicom/DownloadUtil.java
	src/main/java/edu/stanford/epad/epadws/service/PluginOperations.java

+++ src/main/java/edu/stanford/epad/epadws/models/WorkListToStudy.java
@@ -137,7 +137,7 @@ public class WorkListToStudy extends AbstractDAO {
         {"workListId","long","worklist_id","integer"},
         {"studyId","long","study_id","integer"},
         {"projectId","long","project_id","integer"},
-        {"sortOrder","int","sortorder","integer"},
+        {"sortOrder","long","sortorder","integer"},
         {"startDate","Date","startdate","date"},
         {"completeDate","Date","completedate","date"},
         {"status","String","status","varchar"},
